Move SummaryState implementation to EclipseState/Schedule/
