+++ opm/output/eclipse/Summary.hpp
@@ -27,10 +27,9 @@
 #include <ert/ecl/ecl_sum.h>
 
 #include <opm/parser/eclipse/EclipseState/Grid/EclipseGrid.hpp>
+#include <opm/parser/eclipse/EclipseState/Schedule/SummaryState.hpp>
 
 #include <opm/output/data/Wells.hpp>
-
-#include <opm/output/eclipse/SummaryState.hpp>
 #include <opm/output/eclipse/RegionCache.hpp>
 
 namespace Opm {
